assets/2019-11-30/requirements.txt
-i https://pypi.org/simple
certifi==2019.11.28
chardet==3.0.4
click==7.0
flask-opentracing==1.1.0
flask==1.1.1
grpcio-opentracing==1.1.4
grpcio-tools==1.25.0
grpcio==1.25.0
idna==2.8
invoke==1.3.0
itsdangerous==1.1.0
jaeger-client==4.2.0
jinja2==2.10.3
markupsafe==1.1.1
opentracing==2.2.0
protobuf==3.11.0
requests==2.22.0
six==1.13.0
threadloop==1.0.2
thrift==0.13.0
tornado==6.0.3
urllib3==1.25.7
werkzeug==0.16.0
